+++ StarSaver/Constants.h
@@ -1,0 +1,14 @@
+//  Constants.h
+//  StarSaver
+
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+extern NSString * const kModuleName;
+
+extern NSString * const  kNumberOfStars;
+extern NSString * const  kNovaProbability;
+extern NSString * const  kAnimationTiming;
+
+NS_ASSUME_NONNULL_END