+++ StarSaver/Constants.m
@@ -1,0 +1,10 @@
+//  Constants.m
+//  StarSaver
+
+#import "Constants.h"
+
+NSString * const kModuleName = @"com.vinorodrigues.StarSaver";
+
+NSString * const kNumberOfStars = @"numberOfStars";
+NSString * const kNovaProbability = @"novaProbability";
+NSString * const kAnimationTiming = @"animationTiming";