+++ StarSaver/StarSaverView.m
@@ -8,6 +8,8 @@
  */
 
 #import "StarSaverView.h"
+#import "ConfigureSheetController.h"
+#import "Constants.h"
 #import <Foundation/Foundation.h>
 #import <AppKit/AppKit.h>
 #import <math.h>
@@ -15,13 +17,43 @@
 @implementation Star
 @end
 
+@interface StarSaverView ()
+
+  // Private vars
+  @property (nonatomic, assign) CGFloat width;   // Width of screen at `init`
+  @property (nonatomic, assign) CGFloat height;  // Height of screen at `init`
+  @property (nonatomic, assign) NSInteger cols;  // Width / {star resources width}
+  @property (nonatomic, assign) NSInteger rows;  // Height / {star resources height}
+  @property (nonatomic, assign) BOOL doOffsets;  // To offset or not
+
+  @property (strong) ConfigureSheetController *configureSheetController;
+
+  @property (assign) NSInteger numberOfStars;    // Number of stars
+  @property (assign) NSInteger novaProbability;  // Probability for Nova (1 in X chance)
+  @property (assign) NSInteger animationTiming;  // Animation timing in milliseconds
+
+  // Private methods
+  - (NSPoint)randomPosition;
+  - (NSPoint)randomOffset;
+  - (NSRect)getStarRect:(Star*)star;
+  - (void)internalInit;
+  - (BOOL)isMiniPreview;
+@end
+
 @implementation StarSaverView
+
+// ==================================================
+#pragma mark - Init Methods
+// ==================================================
 
 // ------------------------------
 // Initializer for normal and preview mode
 // ------------------------------
 - (instancetype)initWithFrame:(NSRect)frame isPreview:(BOOL)isPreview {
   self = [super initWithFrame:frame isPreview:isPreview];
+  #ifdef DEBUG
+  NSLog(@"StarSaverView initWithFrame:isPreview:%@", isPreview ? @"YES" : @"NO");
+  #endif
   if (self) {
     [self internalInit];
   }
@@ -33,6 +65,9 @@
 // ------------------------------
 - (instancetype)initWithCoder:(NSCoder *)coder {
   self = [super initWithCoder:coder];
+  #ifdef DEBUG
+  NSLog(@"StarSaverView initWithCoder:");
+  #endif
   if (self) {
     [self internalInit];
   }
@@ -40,31 +75,105 @@
 }
 
 // ------------------------------
-// Consolidated initialisation methods
+// Consolidated initialisation method
 // ------------------------------
 - (void)internalInit {
+  #ifdef DEBUG
+  NSLog(@"StarSaverView internalInit");
+  #endif
+  self.isRunning = false;
+  
   // ----- Initialize RandomSeed -----
+  #ifdef DEBUG
+  NSLog(@"StarSaverView internalInit A");
+  #endif
   // Use the current time to set a unique seed
   srand((unsigned int)time(NULL));  // used by `rand()`
   srandom((unsigned int)time(NULL));  // used by `random()`
 
   // ----- Get the size of the screen -----
+  #ifdef DEBUG
+  NSLog(@"StarSaverView internalInit B");
+  #endif
   self.width = self.bounds.size.width; if (self.width <= 0) { self.width = 1; }
   self.height = self.bounds.size.height; if (self.height <= 0) { self.width = 1; }
   self.cols = floor( self.width / STAR_CELL_WIDTH ); if (self.cols <= 0) { self.cols = 1; }
   self.rows = floor( self.height / STAR_CELL_HEIGHT ); if (self.rows <= 0) { self.rows = 1; }
-
+  #ifdef DEBUG
+  NSLog(@"StarSaverView internalInit self.width = %ld, self.height = %ld", (long)self.width, (long)self.height);
+  NSLog(@"StarSaverView internalInit self.cols = %ld, self.rows = %ld", (long)self.cols, (long)self.rows);
+  #endif
+  
   // ----- Preferences -----
   
-  // Load user configuration
-  [self loadPreferences];
-  
-  // Adjust star count and animation timing based on screen width
-  if (self.isPreview && self.width <= 639) {
-    self.numberOfStars = 5;      // Set to 5 stars for smaller screens
-    self.animationTiming = 2000; // Set to 2 seconds for smaller screens
+  #ifdef DEBUG
+  NSLog(@"StarSaverView internalInit C");
+  #endif
+  NSNumber *fallbackNumberOfStars;
+  NSNumber *fallbackNovaProbability;
+  NSNumber *fallbackAnimationTiming;
+
+  // Default to 1% of the cell count
+  // On Legacy iMac 27" that's `((2560/14)*(1440/14))*0.01` = `188`
+  // Have not tested this on a Retina display :(
+  fallbackNumberOfStars = @(floor((self.rows * self.cols) * 0.01));
+  // Default Nova probability to 1 in 25
+  fallbackNovaProbability = @(25);
+  // Default animation timing set so that each star lives for aprox. 1 m
+  if (self.numberOfStars > 0) {
+    fallbackAnimationTiming = @(floor((1000 * 60) / ((int)fallbackNumberOfStars / 2)));
+  } else {
+    fallbackAnimationTiming = @(250);
   }
-    
+
+  #ifdef DEBUG
+  NSLog(@"StarSaverView internalInit 1");
+  #endif
+  ScreenSaverDefaults *defaults = [ScreenSaverDefaults defaultsForModuleWithName:kModuleName];
+
+  #ifdef DEBUG
+  NSLog(@"StarSaverView internalInit D");
+  #endif
+  // Register default preferences
+  [defaults registerDefaults:@{
+    kNumberOfStars: fallbackNumberOfStars,
+    kNovaProbability: fallbackNovaProbability,
+    kAnimationTiming: fallbackAnimationTiming
+    }];
+
+  // ----- Set up self vars -----
+
+  // Load user configuration
+  #ifdef DEBUG
+  NSLog(@"StarSaverView internalInit E");
+  #endif
+  [self loadPreferences];
+
+  if (!self.isMiniPreview) {
+    #ifdef DEBUG
+    NSLog(@"StarSaverView internalInit E.1");
+    #endif
+    self.starHead = floor( self.numberOfStars / 4);
+  } else {
+    #ifdef DEBUG
+    NSLog(@"StarSaverView internalInit E.2");
+    #endif
+    self.starHead = self.numberOfStars - 1; // Set to end of star array
+  }
+  
+  #ifdef DEBUG
+  NSLog(@"StarSaverView internalInit self. .. numberOfStars = %ld, novaProbability = %ld, animationTiming = %ld, starHead = %ld",
+        (long)self.numberOfStars,
+        (long)self.novaProbability,
+        (long)self.animationTiming,
+        (long)self.starHead );
+  #endif
+  
+  // ----- Load Images -----
+  
+  #ifdef DEBUG
+  NSLog(@"StarSaverView internalInit F");
+  #endif
   // Load all the star images (`star1.png` to `star5.png`)
   NSBundle *bundle = [NSBundle bundleForClass:[self class]];
   NSMutableArray *loadedImages = [NSMutableArray array];
@@ -80,10 +189,10 @@
   self.starImages = [loadedImages copy];
   
   // ----- Initialise the Stars -----
-  
-  // self.starHead = 0;  // Start the star "buffer" at the begining
-  self.starHead = floor( self.numberOfStars / 4);  // Initially show 1/4 the stars ;)  OG started at 0, but...
-
+ 
+  #ifdef DEBUG
+  NSLog(@"StarSaverView internalInit G");
+  #endif
   // Initialize the stars array
   self.stars = [NSMutableArray array];
 
@@ -92,65 +201,60 @@
     Star *star = [[Star alloc] init];
     if (i >= self.starHead) {
       star.state = StarStateGone;
-      // star.position = NSMakePoint(0, 0); // not needed
     } else {
       star.state = StarStateNormal;
       star.position = self.randomPosition;
-      star.offs = self.randomOffs;
+      if (self.doOffsets) {
+        star.offset = self.randomOffset;
+      } else {
+        star.offset = NSMakePoint(0, 0);
+      }
     }
     [self.stars addObject:star];
   }
 
-  //  // Needs `animateOneFrame`
-  //  NSTimeInterval interval = self.animationTiming / 1000.0;  // settings is in ms
-  //  [self setAnimationTimeInterval:interval];
+  #ifdef DEBUG
+  NSLog(@"StarSaverView internalInit H");
+  #endif
+  // Needs `animateOneFrame`
+  NSTimeInterval interval = self.animationTiming / 1000.0;  // settings is in ms
+  [self setAnimationTimeInterval:interval];
 }
+
+// ==================================================
+#pragma mark - Preference Methods
+// ==================================================
 
 // ------------------------------
 // Load the screen saver settings
 // ------------------------------
-// TODO : Change this to use the `ScreenSaverDefaults` class
 - (void)loadPreferences {
-  // Load configuration from the JSON file
-  NSString *configFilePath = [[NSBundle bundleForClass:[self class]] pathForResource:@"config" ofType:@"json"];
+  #ifdef DEBUG
+  NSLog(@"StarSaverView loadPreferences");
+  #endif
   
-  if (configFilePath) {
-    NSData *data = [NSData dataWithContentsOfFile:configFilePath];
-    NSError *error;
-    NSDictionary *json = [NSJSONSerialization JSONObjectWithData:data options:0 error:&error];
-      
-    if (!error && json) {
-      NSNumber *starCount = json[@"starCount"];
-      NSNumber *novaProbability = json[@"novaProbability"];
-      NSNumber *animationTiming = json[@"animationTiming"];
-          
-      if (starCount) {
-        self.numberOfStars = [starCount integerValue];
-      }
-      if (novaProbability) {
-        self.novaProbability = [novaProbability integerValue];
-      }
-      if (animationTiming) {
-        self.animationTiming = [animationTiming integerValue];
-      }
-    }
-  }
+  if (!self.isMiniPreview) {
+    ScreenSaverDefaults *defaults = [ScreenSaverDefaults defaultsForModuleWithName:kModuleName];
 
-  // Set default values if config is invalid
-  if ((self.numberOfStars <= 0) || (self.numberOfStars > 1000)) {  // 1 to 1000
-    // Default to 1% of the cell count
-    // On Legacy iMac 27" that's `((2560/14)*(1440/14))*0.01` = `188`
-    // Have not tested this on a Retina display :(
-    self.numberOfStars = floor((self.rows * self.cols) * 0.01);
+    self.numberOfStars = [defaults integerForKey:kNumberOfStars];
+    self.novaProbability = [defaults integerForKey:kNovaProbability];
+    self.animationTiming = [defaults integerForKey:kAnimationTiming];
+  } else {
+    self.numberOfStars = 10;     // Set to 5 stars for smaller screens
+    self.novaProbability = 25;   // 1 in 25 probability
+    self.animationTiming = 2000; // Set to 2 seconds for smaller screens
   }
-  if ((self.novaProbability < 0) || (self.novaProbability > 1000)) {  // 0 to 1000, NB: 0 = always
-    // Default Nova probability to 1 in 25
-    self.novaProbability = 25;
-  }
-  if ((self.animationTiming < 60) || (self.animationTiming > 60000)) {  // 1/60 s to 1 min
-    // Default animation timing set so that each star lives for aprox. 1 m
-    self.animationTiming = floor((1000 * 60) / (self.numberOfStars / 2));
-  }
+}
+
+// ==================================================
+#pragma mark - Private Helper Methods
+// ==================================================
+
+// ------------------------------
+// Assess if in System Preferences mini view
+// ------------------------------
+- (BOOL)isMiniPreview {
+  return self.isPreview && (self.bounds.size.width < 640);
 }
 
 // ------------------------------
@@ -164,7 +268,7 @@
 // ------------------------------
 // Generate random offsets
 // ------------------------------
-- (NSPoint)randomOffs {
+- (NSPoint)randomOffset {
   return NSMakePoint( SSRandomIntBetween(0, (int)STAR_CELL_WIDTH),
                       SSRandomIntBetween(0, (int)STAR_CELL_HEIGHT) );
 }
@@ -173,8 +277,8 @@
 // Get the Rect that the star lives in
 // ------------------------------
 - (NSRect)getStarRect:(Star *)star {
-  NSInteger ox = star.offs.x;
-  NSInteger oy = star.offs.y;
+  NSInteger ox = star.offset.x;
+  NSInteger oy = star.offset.y;
 
   // adjust so that it's not off screen
   while (((star.position.x * STAR_CELL_WIDTH) + ox) > self.width) {
@@ -191,34 +295,13 @@
 }
 
 // ------------------------------
-- (void)startAnimation {
-  [super startAnimation];
-  
-  // Create a timer that fires {every X} and calls the 'timerTick' method
-
-  NSTimeInterval interval = self.animationTiming / 1000.0;
-  self.timer = [NSTimer scheduledTimerWithTimeInterval:interval
-                                                target:self
-                                              selector:@selector(timerTick)
-                                              userInfo:nil
-                                               repeats:YES];
-
-  [self setNeedsDisplay:YES];  // redraw the whole screen
-}
-
-// ------------------------------
-- (void)stopAnimation {
-  [super stopAnimation];
-
-  // Invalidate the timer when stopping the screen saver
-  [self.timer invalidate];
-  self.timer = nil;
-}
-
-// ------------------------------
 // Draw star at n-th position
 // ------------------------------
 - (void)drawStarAt:(NSInteger)index {
+  #ifdef DEBUG
+  NSLog(@"StarSaverView drawStarAt:%ld", (long)index);
+  #endif
+
   NSImage *starImage;
   Star *star = [self.stars objectAtIndex:(index)];  // Get the start at the head position
 
@@ -250,11 +333,46 @@
   }
 }
 
+// ==================================================
+#pragma mark - Key Screen Saver Methods
+// ==================================================
+
+// ------------------------------
+- (void)startAnimation {
+  #ifdef DEBUG
+  NSLog(@"StarSaverView startAnimation");
+  #endif
+  [super startAnimation];
+ 
+  [self loadPreferences];
+  
+  self.isRunning = YES;
+  [self setNeedsDisplay:YES];  // redraw the whole screen
+}
+
+// ------------------------------
+- (void)stopAnimation {
+  #ifdef DEBUG
+  NSLog(@"StarSaverView stopAnimation");
+  #endif
+  [super stopAnimation];
+  
+  self.isRunning = NO;
+}
+
 // ------------------------------
 // Draw area needed
 // ------------------------------
 - (void)drawRect:(NSRect)rect {
   [super drawRect:rect];
+  
+  #ifdef DEBUG
+  NSLog(@"StarSaverView drawRect (%ld, %ld, %ld, %ld)",
+        (long)rect.origin.x,
+        (long)rect.origin.y,
+        (long)rect.size.width,
+        (long)rect.size.height );
+  #endif
   
   // Fill the background with black
   [[NSColor blackColor] setFill];
@@ -269,10 +387,30 @@
   }
 }
 
+// ------------------------------
+// Gets called repeatedly to draw states on timer ticks
+// when used with `setAnimationTimeInterval`
+// ------------------------------
+- (void)animateOneFrame {
+  [super animateOneFrame];
+  
+  #ifdef DEBUG
+  NSLog(@"StarSaverView animateOneFrame");
+  #endif
+  
+  if (self.isRunning) {
+    [self timerTick];
+  }
+  return;
+}
+
 // ---------------------------------
 // Abstracting Timer Ticks here
 // ---------------------------------
 - (void)timerTick {
+  #ifdef DEBUG
+  NSLog(@"StarSaverView timerTick (%ld)", (long)self.starHead);
+  #endif
   
   NSInteger index = self.starHead;
   
@@ -306,7 +444,9 @@
 
       // New position
       star.position = self.randomPosition;  // new position
-      star.offs = self.randomOffs;  // new offset
+      if (self.doOffsets) {
+        star.offset = self.randomOffset;  // new offset
+      }
   
       // inc the header, i.e. move on to the next star
       self.starHead++;
@@ -322,26 +462,37 @@
   [self setNeedsDisplayInRect:[self getStarRect:star]];
 }
 
-//// ------------------------------
-//// Gets called repeatedly to draw states on timer ticks 
-//// when used with `setAnimationTimeInterval`
-//// ------------------------------
-//- (void)animateOneFrame {
-//  [super animateOneFrame];
-//  [self timerTick];
-//  return;
-//}
+// ==================================================
+#pragma mark - Configuration Sheet Methods
+// ==================================================
 
 // ------------------------------
 // ------------------------------
 - (BOOL)hasConfigureSheet {
-  return NO;
+  #ifdef DEBUG
+  NSLog(@"StarSaverView hasConfigureSheet");
+  #endif
+  return YES;
 }
 
 // ------------------------------
 // ------------------------------
 - (NSWindow*)configureSheet {
-  return nil;
+  #ifdef DEBUG
+  NSLog(@"StarSaverView configureSheet");
+  #endif
+  
+  if (!self.configureSheetController) {
+    self.configureSheetController = [[ConfigureSheetController alloc] init];
+  }
+  #ifdef DEBUG
+  if (self.configureSheetController.window == nil) {
+    NSLog(@"Error: configureSheetController.window is nil");
+  } else {
+    NSLog(@"Returning configureSheetController.window: %@", self.configureSheetController.window);
+  }
+  #endif
+  return self.configureSheetController.window;
 }
 
 @end